+++ src/Console/Exception/InfectionException.php
@@ -4,11 +4,14 @@
  *
  * License: https://opensource.org/licenses/BSD-3-Clause New BSD License
  */
-
 declare(strict_types=1);
 
 namespace Infection\Console\Exception;
 
 class InfectionException extends \Exception
 {
+    public static function configurationAborted()
+    {
+        return new self('Configuration aborted');
+    }
 }