+++ src/Process/Listener/FileLoggerSubscriber/TextFileLogger.php
@@ -4,63 +4,17 @@
  *
  * License: https://opensource.org/licenses/BSD-3-Clause New BSD License
  */
-
 declare(strict_types=1);
 
-namespace Infection\Process\Listener;
+namespace Infection\Process\Listener\FileLoggerSubscriber;
 
-use Infection\Config\InfectionConfig;
-use Infection\Console\LogVerbosity;
-use Infection\EventDispatcher\EventSubscriberInterface;
-use Infection\Events\MutationTestingFinished;
-use Infection\Filesystem\Filesystem;
-use Infection\Mutant\MetricsCalculator;
 use Infection\Process\MutantProcess;
 
-class TextFileLoggerSubscriber implements EventSubscriberInterface
+class TextFileLogger extends FileLogger
 {
-    /**
-     * @var InfectionConfig
-     */
-    private $infectionConfig;
-
-    /**
-     * @var MetricsCalculator
-     */
-    private $metricsCalculator;
-
-    /**
-     * @var Filesystem
-     */
-    private $fs;
-
-    /**
-     * @var bool
-     */
-    private $isDebugMode;
-
-    public function __construct(
-        InfectionConfig $infectionConfig,
-        MetricsCalculator $metricsCalculator,
-        Filesystem $fs,
-        int $logVerbosity = LogVerbosity::DEBUG
-    ) {
-        $this->infectionConfig = $infectionConfig;
-        $this->metricsCalculator = $metricsCalculator;
-        $this->fs = $fs;
-        $this->isDebugMode = ($logVerbosity === LogVerbosity::DEBUG);
-    }
-
-    public function getSubscribedEvents()
+    public function writeToFile()
     {
-        return [
-            MutationTestingFinished::class => [$this, 'onMutationTestingFinished'],
-        ];
-    }
-
-    public function onMutationTestingFinished(MutationTestingFinished $event)
-    {
-        $logFilePath = $this->infectionConfig->getTextFileLogPath();
+        $logFilePath = $this->infectionConfig->getLogPathInfoFor('text');
 
         if ($logFilePath) {
             $logs[] = $this->getLogParts($this->metricsCalculator->getEscapedMutantProcesses(), 'Escaped');
@@ -73,13 +27,7 @@ class TextFileLoggerSubscriber implements EventSubscriberInterface
 
             $logs[] = $this->getLogParts($this->metricsCalculator->getNotCoveredMutantProcesses(), 'Not covered');
 
-            $this->fs->dumpFile(
-                $logFilePath,
-                implode(
-                    array_merge(...$logs),
-                    "\n"
-                )
-            );
+            $this->write($logs, $logFilePath);
         }
     }
 
@@ -87,9 +35,9 @@ class TextFileLoggerSubscriber implements EventSubscriberInterface
      * @param MutantProcess[] $processes
      * @param string $headlinePrefix
      *
-     * @return array
+     * @return string
      */
-    private function getLogParts(array $processes, string $headlinePrefix): array
+    private function getLogParts(array $processes, string $headlinePrefix): string
     {
         $logParts = $this->getHeadlineParts($headlinePrefix);
 
@@ -106,7 +54,7 @@ class TextFileLoggerSubscriber implements EventSubscriberInterface
             }
         }
 
-        return $logParts;
+        return implode($logParts, "\n");
     }
 
     private function getHeadlineParts(string $headlinePrefix): array