+++ tests/Php/XdebugHandlerTest.php
@@ -4,7 +4,6 @@
  *
  * License: https://opensource.org/licenses/BSD-3-Clause New BSD License
  */
-
 declare(strict_types=1);
 
 namespace Infection\Tests\Php;
@@ -47,6 +46,10 @@ class XdebugHandlerTest extends TestCase
 
     protected function setUp()
     {
+        if (PHP_SAPI === 'phpdbg') {
+            $this->markTestSkipped();
+        }
+
         putenv(PhpIniHelper::ENV_ORIGINALS_PHP_INIS);
         putenv(XdebugHandlerMock::ENV_DISABLE_XDEBUG);
         putenv(ConfigBuilder::ENV_TEMP_PHP_CONFIG_PATH);