+++ tests/Process/ExecutableFinder/PhpExecutableFinderTest.php
@@ -4,11 +4,11 @@
  *
  * License: https://opensource.org/licenses/BSD-3-Clause New BSD License
  */
-
 declare(strict_types=1);
 
 namespace Infection\Tests\Process\ExecutableFinder;
 
+use Infection\Filesystem\Filesystem;
 use Infection\Php\ConfigBuilder;
 use Infection\Process\ExecutableFinder\PhpExecutableFinder;
 use PHPUnit\Framework\TestCase;
@@ -43,6 +43,6 @@ class PhpExecutableFinderTest extends TestCase
 
     public function tearDown()
     {
-        @unlink($this->workspace);
+        (new Filesystem())->remove($this->workspace);
     }
 }